Rework public configurations to support multiple index files.
Previously, the configuration only supported building individual
files, but for more complex sites, this may prove to be a blocker
to using this plugin. To work around this, this commit refactors
the public API to take an array of individual configuration
objects. This also moves us away from requiring wholesale
GraphQL queries and instead moves to using individual resolvers
functions.